Improve advisory lock module.
Use noop for Android and iOS as well as wasm32-unknown-unknown.
